+++ src/mlpack/methods/cf/normalization/z_score_normalization.hpp
@@ -79,8 +79,8 @@ class ZScoreNormalization
    * @param item Item ID.
    * @param rating Computed rating before denormalization.
    */
-  double Denormalize(const int /* user */,
-                     const int /* item */,
+  double Denormalize(const size_t /* user */,
+                     const size_t /* item */,
                      const double rating) const
   {
     return rating * stddev + mean;
